rename service suffix from services that doesnt have class and homogenize to be singular
