+++ admin-dev/themes/new-theme/js/pages/product/combination/combination-list/bulk-edition-handler.ts
@@ -26,7 +26,7 @@
 import {FormIframeModal} from '@components/modal';
 import ProductMap from '@pages/product/product-map';
 import ProductEvents from '@pages/product/product-event-map';
-import CombinationsService from '@pages/product/service/combinations-service';
+import CombinationService from '@pages/product/service/combination-service';
 import {EventEmitter} from 'events';
 import BulkChoicesSelector from '@pages/product/combination/combination-list/bulk-choices-selector';
 import ProgressModal from '@components/modal/progress-modal';
@@ -40,7 +40,7 @@ const CombinationEvents = ProductEvents.combinations;
 export default class BulkEditionHandler {
   readonly productId: number;
 
-  private combinationsService: CombinationsService;
+  private combinationsService: CombinationService;
 
   private eventEmitter: EventEmitter;
 
@@ -52,7 +52,7 @@ export default class BulkEditionHandler {
     productId: number,
     eventEmitter: EventEmitter,
     bulkChoicesSelector: BulkChoicesSelector,
-    combinationsService: CombinationsService,
+    combinationsService: CombinationService,
   ) {
     this.productId = productId;
     this.eventEmitter = eventEmitter;
