+++ admin-dev/themes/new-theme/js/pages/product/combination/combination-list/row-delete-handler.ts
@@ -26,7 +26,7 @@
 import ProductMap from '@pages/product/product-map';
 import ProductEventMap from '@pages/product/product-event-map';
 import {EventEmitter} from 'events';
-import CombinationsService from '@pages/product/service/combinations-service';
+import CombinationService from '@pages/product/service/combination-service';
 import ConfirmModal from '@components/modal/confirm-modal';
 
 const {$} = window;
@@ -39,11 +39,11 @@ const CombinationEvents = ProductEventMap.combinations;
 export default class RowDeleteHandler {
   private eventEmitter: EventEmitter;
 
-  private combinationsService: CombinationsService;
+  private combinationsService: CombinationService;
 
   constructor(
     eventEmitter: EventEmitter,
-    combinationsService: CombinationsService,
+    combinationsService: CombinationService,
   ) {
     this.eventEmitter = eventEmitter;
     this.combinationsService = combinationsService;
