+++ admin-dev/themes/new-theme/js/pages/product/service/attachment.ts
