+++ admin-dev/themes/new-theme/js/pages/product/service/attribute-group.ts
