+++ admin-dev/themes/new-theme/js/pages/product/service/image.ts
